geo_bath.i:  Modified function raspulsearch to plot the graph from function ex_veg while examining a point processed for Vegetation.
veg.i:  Modified several functions in veg.i to define the peak of the last return as the bottom elevation.  As of now, if 'Correct Range Walk with Centroid' is selected in the 'Process EAARL Data' GUI, the 'peak algorithm' is used to find the bare earth.  This 'peak algorithm' checks for saturation of the return signal and uses channels 2 or 3 if needed.  The algorithm also rejects peaks that are formed with a leading + trailing edge of less than 4 nanoseconds.  These are assumed to be noise events. Instead of assuming that the peak will occur within, say, 12ns from the start of the leading edge, the algorithm finds the 'distance' from the start of the leading edge to the end of the trailing edge (as defined by variable 'retdist') and finds the peak of the signal within that range.

If 'Correct Range Walk with Centroid' is not selected in 'Process EAARL Data GUI', the algorithm for vegetation will not do any of the stuff I explained above, and will work exactly like it did earlier.