comparison_fns.i:  Modified function compare_pts to make certain variables external to the function.
veg.i:  Made several changes to the vegetation algorithm for bald earth mapping.  The new algorithm uses the irange value to compute the last return.  It also uses the centroid of the transmit pulse and the centroid of the first return.  Basic trigonometry using similar triangles was used to find the bald earth elevations.  This algorithm currently uses the peak of the bottom return.
   The output structure of the new bald earth algorithm has been replaced with VEG_ALL and VEG_.  Structures VEGALL and VEG were left unchanged to offer compatibility with the old veg algorithm.

read_yfile.i:  Modified to read in new structures for the new veg algorithm.

geo_bath.i:  Modified to include new veg structures in function sel_data_rgn().