comparison_fns.i:  Modified function read_txt_anal_file to calculate the  number of lines in the input text file.
veg.i:  Modified function make_fs_veg to include the change in the x and y values along the range vector that defines bare earth.  This is an important change!  As of now, the output structure has only one set of easting, northing values for each waveform.  I have temporarily set these easting and northing values to define the bare earth location.  We may have to change this structure format to include the easting northing values of the first surface (canopy) location.
       I have also begun writing a function to extract the height and location of *ALL* peaks in the waveform.  This function will remain 'under construction' for some time but will not affect any regular processing routines.