batchmaker.pro: Batchmaker will automatically generate the batchfile required in batch_process. To run, input the directory to save all data to (e.g. ~/Data/EAARl/TampaBay) and the UTM easting/northing of the northwest corner of the index tile. You can optionally set to define a single-character tag that describes the type of data to be processed (e.g. "b" for bathy or "v" for veg). Batchmaker creates the batchfile and all directories needed to hold the to be processed data.
multibatch.pro:
Multibatch is used to create a rectangular grid of index tiles. It works basically the
same as bachmaker (it calls batchmaker). To run, input the UTM easting/northing of the
northwest corner of the northwest index tile and also the number of index tiles to creat
east and south (the dimensions of the rectangle in units of index tile). You may also
optionally define a tag described above.