*** Important Update ***
 The previous version of eaarl_mounting_bias.i had the wrong offset from the
antenna to the mirror, so if you processed anything with it, you will need
to re-run it with this corrected file.  The previous version would introduce
a 1.3 meter offset in the lidar elevations.  The problem was the correct
z-axis offset was in x-axis, and the z-axis (up/down) was set to zero.