Significant bug corrected which caused a 1-meter "deep bias" in bathy data.  The problem was that we were measuring the depth by measuring the time from the leading edge of the digitized waveform instead of the centroid of the surface return.  The "deep bias" was almost exactly 1-meter.  It was/is not present in the waveform based depth displays, but was in the numeric depth computations only.
In order to address the problem several files were modified and a new parameter
was added to type "R" structures to hold the time interval from the leading edge
of the waveform to the centroid of the surface return.  The data in that array
is then subtracted from the the depth values resulting in the actual time from the
surface return to the bottom peak.  No corrections have been made (yet) for AOI of
the surface and the effect seems to be minimal for submarine topographic mapping
anyway.

All time-interval based bathy type data should be re-processed.