Subtracted the mean of the good elevations instead of using the median of them all.
Added auto-scaling to the topography display gray-scale limits to +/- 3 sigma

Replace the medians of all the x,y values with the means of the good x,y values

Commented out reversing every other scan line to make all go in same direction
because it was unnecessary when simply binning the good values.  If that is
ever reimplemented is should be done before any editing.