triangulate.i: Added a function in yorick that does a Delaunay triangulation on an x,y,z point cloud.  The function reads in an x,y,z ascii file or a processed data array (currently works only for depth_all (bathymetry)) and returns the list of the indices to the vertices that form the triangles.  This function is adapted from a C program. The resulting triangulated mesh can be plotted (if plot=1) in window, 0.
The triangulation function in yorick is *SLOW*.  It takes a few seconds to
triangulate 1-2K points, but it took 10minutes to triangulate 21500 points.
The C version of the function is probably 100 times faster, but I found a
better C program (that uses a divide-and-conquer approach) which we need to
eventually link with yorick.

I envision using this function (before the C version is linked) to filter
data -- during manual and automated (rcf) filtering.