batch_process.i: Modified function batch_rcf: 	-Added keyword 'interactive='. If set to 1 and rcfmode=2 the program will call rcf_triag_filter with interactive=1, plottriag=1, and datawin=5. It will also plot the raw data of the current data tile in window 5. 	-Added keyword 'onlyupdate='. If set to 1 it will check to see if an RCF file exists for a region. If it does, the program will skip to the next region. This allows one to resume a crashed batch_rcf process without manually changing the for loop to start where the program crashed.
ytriangulate.i: Modified function triangulate_xyz to check for the presence of a lock file before proceeding with the triangulation. If the lock file does not exist, the program runs the triangulation and removes the lockfile. If it does exist, it waits until the lock file is deleted. This prevents a crash when two people using the same system attempt to triangulate and use the same temp file.

ircf.i: Modified function rcf_triag_filter to skip over a region entierly if 's' is entered at the 'yes/no/done/end' prompt. This is saves time when most of a region is undesired data.