Modified functions sdist() and mdist() to allow the user to work in UTM coordinates. This will allow the flight planning to work in UTM or latlon and can also be used in planning ground survey transects.
Added 2 optional keywords to sdist(): in_utm and out_utm.
When in_utm is set, it is assumed the user is working in UTM coordinates.
The flight plan will be plotted in UTM but written out in the regular latlon
format.
If out_utm is set, then the output format changes to:
utmseg 19-1 e485227.88:n4226044.82 e485869.57:n4225799.03

Also modified function mdist() so that the return array when working in
UTM coordinates is similar to when working in latlon.