Added files confload.ytk and confload.i: 	Added configuration file loader. Set up the conf file once for each mission day and then load missions (all files and settings) in a single click. Configuration files must be generated by hand and placed in the root folder of the mission day. Should also be named mm-dd-yy.conf. Here's an example conf file with (comments):
/data/0/tampa_bay_04/03-03-04/eaarl/040303-edb.idx (location of edb file)
/data/0/tampa_bay_04/03-03-04/gps-063/3-3-04-cfcg-n111x-p-wgs84-6-7-04-ver-cmb-pnav.ybin (location of pnav file)
/data/0/tampa_bay_04/03-03-04/dmars/3-3-04-cfcg-n111x-p-wgs84-6-7-04-ver-fwd-ins.pbd (location on attitude file. Can use either dmars .pbd or tans .ybin files)
~/lidar-processing/maps/fla.pbd (location of coastline map file)
313099,371543,3.0508e+06,3.10924e+06 (limits for region of interest)
/data/0/tampa_bay_04/03-03-04/dmars/03-03-04-ops_conf.i (Location of command file to initialize ops_conf)
bays (type of bathy settings. Set to 'veg' to skip. If this is set to 'bays' 'clear' 'crappy' or 'super shallow' it loads those presets. Otherwise it #includes the line as a command file)
none (This line (unless it is 'none') is #included to load any other functions or run any other commands for the dataset. Could set as a command file that pops up an important message about the mission)

To use the program load ytk and eaarl.ytk. Then click file -> Load EAARL from Conf. In the window that opens, click the browse button and select the root directory of the mission (e.g. /data/0/tampa_bay_04/ in the example above). Click get confs and a select-box will appear. Click the mission you want to load and click 'Load conf'. The name of the file is then passed to the yorick config file loader (in confload.i).

Modified file eaarl.ytk:
	Added Load EAARL from Conf menu option.