Added Files: mktrajfiles
After using a combination of 'find' and 'rsync' to get updated
trajectories from grafnav, all of the zip files get put into a
single directory at the rsync destination.  This then resulted
in manually moving or copying the files into the proper mission
subdirectory.  If the files were "moved", then after the next
rsync, one had to figure out which files were really new.

By rsync'ing the files into a YYYY/trajectories directory first,
mktrajfiles then looks into the .. (YYYY) directory and automatically
creates the necessary subdirectories and unzips the files where
they belong.  It also runs pnav2bin on the gps files.  If the
destination already exists, then it skips that file and goes to
the next one.

Subsequent rsyncs from grafnav are now more efficient (because the
zip files are left in YYYY/trajectories) and all of the follow-up
work is done by mktrajfiles.

Now if we just had a way to run iex_ascii2pbd from the unix shell.