Improved the plot_colorbar procedure so that the colorbar is plotted at any location defined by the user in a window.  The x0,y0 location for the colorbar is defined in normalized screen coordinates (xnorm, ynorm).  The size of the window can also be modified (xsize,ysize).
Modified procedure make_GE_plots so that the colorbar is generated for each 2k by 2k datatile.  The colorbar is saved in a file name *_cb.png. The colorbar can then be embedded in the balloon box for each tile image plotted in Google Earth.

The GE images (and colorbar) can be saved as gifs, pngs, tiffs, or jpegs.  Only the gifs and pngs support transparency.