Added option in 'eaarl' toolbar under Images->Load Images in Yorick...
*******************************
This functionality will not work without the yorick-z plugin.  The yorick-z plugin is available for download at http://www.maumae.net/yorick/doc/plugins.php
*******************************

When this option is selected, a GUI is invoked which allows the user to load any image (jpg,png,tiff,gif, or other) in Yorick.
If an image world file (jgw, pgw, tfw, gfw, etc.) is available, it is automatically read to determine the boundary coordinates.  The user also has the option of defining the boundary coordinates by typing them in.
The lidar 2k by 2k tiles are automatically read in based on the location information available in the file name (no need to specify coordinates).

The size of the yorick window can be specified as "Small", "Medium", "Large", or "Huge".

The change_window_size function in change_window_size.i is used to change the size of the window.  Extern (global) variable array  _ytk_window_size contains the window size code for all possible yorick windows. Extern variable _ytk_window_exists is set to 1 when a window is already available.

jpeg_support.i contains several functions to read and plot images.

 -- read_lidar_jpeg(filename=)
   This function reads a jpeg file. This function requires the yorick-z plugin to be installed.  The yorick-z plugin is available for download at http://www.maumae.net/yorick/doc/plugins.php
   If the image to be loaded in is not jpeg, it will convert to jpeg using ImageMagick's convert command.

 -- read_lidar_image_location(filename)
    This function reads the lidar image location from the lidar filename which uses the standard naming convention.
   The standard naming convention for EAARL lidar files is:
       t_e310000_n2726000_17_*.jpg
       where 310000 is the UTM Easting, 2726000 is the UTM Northing, and 17 is the UTM zone number.

 -- plot_lidar_image(img, location, win=, winsize=, dofma=)
    This function plots the 2k by 2k lidar image in window, win.

 -- plot_image(img, location, win=, dofma=, winsize=)
    This function plots an image in window, win.

 -- load_and_plot_image(filename, img=, location=, win=, dofma=, winsize=)
    This function loads and plots an image in window, win.

 -- read_img_world_file(filename)
 	This function read_img_world_file reads the image world file and and returns the boundary locations for the image.