batch_datum_convert.i datum_converter.i nad832navd88.i
Added ability to convert to NAVD88 using the GEOID03 model.
Added new keyword geoid_version:  Set to "geoid99" to use GEOID99 model
or "geoid03" to use GEOID03 model.
If geoid_version is not set, the GEOID03 model will be used.  If the GEOID03
pbd files are not available, the user will be prompted and asked if to continue
with GEOID99 model.
The GEOID03 files are available on lidar.net at:
/mnt/alps/eaarl/tarfiles/geoid03_pbdfiles.tgz
The GEOID99 files are available on lidar.net at:
/mnt/alps/eaarl/tarfiles/geoid99_pbdfiles.tgz

The pbd files must be extracted in the lidar-processing/ subdirectory.

It is recommended that any GEOID conversion in the southern Lousiana area must use the GEOID03 model.
See http://www.ngs.noaa.gov/heightmod/LouisianaControl.shtml