l1pro.ytk
Added the option to use the datum convert function under the post processing
menubar dropdown as well as its current location in the filter tools dropdown.
Also, to elaborate on the previous commit...originially you had to add the variable
manually after it was converted, but now it adds itself to the variable list
automatically.