Attempting to make the "Include Heading" option work in sf_a.tcl. Replaced the "get_heading" proc in sf with a modified copy of cir's. Reworked some of the code for including the heading in show_img.
Headings now work as they do for CIR. There appears to be some lingering
bug though, in that the orientation of a given image may be wrong the first
time you visit it, but if you click back/next once or twice it corrects
itself. This applies to both cir and sf in my testing. (I did not modify
cir at all, so this is a pre-existing problem.)