dir.i:
Added func file_copy. Previously, file copying was typically performed by using
the system command to call the underlying Linux cp command. This was incredibly
ineffecient though because each system call requires a new shell to be spawned,
and that shell is spawned as a duplicate of the current yorick environment
which means that all items in memory go along with it--a very slow process if
you have a lot in memory. See 'help, system' for more details.

The file_copy function uses tkcmd to tell Tcl to copy the file for it if Ytk is
running. If Ytk is not running, then Yorick opens the file in binary mode,
reads it to memory, then writes it to the destination -- which might be
ill-advised for large files. The Tcl method is faster than the Yorick method,
but both are considerably faster than the system method.

cir-mosaic.i:

Modified func gen_cir_tiles.
- I stripped out the toscript= option because it was needlessly confusing
  things. This also fixed a bug that completely prevented the function from
  working when toscript=0.
- I replaced the progress indicator with a call to timer_tick for simpler code.
- I replaced system calls to cp with calls to file_copy, which speeds it up
  noticeably.

ytk and ytk.lite:

The code associated with 'fileevent $ytk_fifo readable' called update at the
end. Use of update can lead to a situation where the event loop is re-entered
recursively. This can lead to a situation of infinite recursion, which results
in nasty errors. Though not encountered often, I was in fact running into this
exact problem sporadically when running a large number of tkcmds in direct
succession. Removing the update call appeared to make it work properly again.

However, removing update is likely not a viable option because then the
interface would freeze on long-running commands. This can be fixed, but would
have required that each of those long-running commands be re-written to use
chains of event bindings, which is not feasible. (See http://wiki.tcl.tk/1255)

Instead, I replaced "update" with "update idletasks". This mitigates the
problem by not processing new events or errors on the call; instead it only
processes idle callbacks and defers the new events and errors until after the
command's execution. This should allow the interface to still refresh, and I
don't think it will interfere with anything. The possibility still exists for
infinite recursion, but this change seems to have stopped the error from
happening in my situation.