info.i: dump_info() for display information on selected pixel
transect.i: mtransect(): added a show=3 to redraw the previous
transect.  This is the only way to get the line back if window 5
has been redrawn and using recall=0.  There should be an array
for all the previous transects, but this just has the last one.

transrch.i:  two new functions, transrch() and mtransrch()
transrch( fs, m, llst): gets mouse input to query a point in
  the transect window.  The point selected will be redrawn
  bigger in both the transect and data plot window.  The raster
  and pixel waveform windows will also be displayed.

mtransch( fs, m, llst): calls transrch() in a loop so that the
  user can easily query multiple points.

For both functions, the user must convert the data to an FS type
and use that when calling mtransect() as the latter returns an
index that is used by the new functions.  A typical invocation
would be:

cln_fs = test_and_clean(fs_all);  // remember to redo this when reprocessing
m=mtransect(cln_fs, _any_other_options_normally_used)
mtransrch,cln_fs, m, llst;

m is returned from mtransect().
llst is an internal variable created by mtranasect().