If having a gui interace to draw a single mtransect is fun, than having buttons to draw multiple mtransects must be great!
Updated interace allows 3* different mtransects to be displayed, each
with their own variables and option settings.  Very quickly replot
mtransects without having to scroll through the cmdline.

Also improved some of the help tips and tried to make it obvious
that the "transrch" and "mtransrch" only works with the data
generated from the first "mtransect" button.

*3: controlled by a single variable, easily increased if more are needed.