scanflatmirror2_direct_vector.i
I needed to understand what this file was doing so I could adapt the math to
another task. In the process, I started documenting it, so that documentation
is the only change made. The code is untouched. I've added a lot of explanatory
comments to the first half of the function. The second half still needs
explanation.