mbatch_process(): added a "now=" option.  If set to 1, process the tiles now instead of creating a cmd file for later.  If set to 0, don't process now, but create cmd files so that each tile can be processed from a server/client setup.
While this makes it essentially the same as batch_process.i, a big block
of code was moved around to turn the actually processing of a tile into
a function and separating it from the code that generates the tiles.

Fixed to only create a tile cmd file if there is data to be processed.

Changed to only draw green boxes around the tile if the area has been
processed (ie: there was a flight line going through the tile).