Server mode now automatically checks every 10 seconds if there are new files to be processed in /tmp/batch/jobs (as placed there by mbatch_process() in ALPS) and then sends them to all connected clients that are "ready".  This removes the need to manually tell each client to start requesting files.
A client will automatically request a new file after completing a file.
This method keeps the turn-around time to a minimum when there are files
in the queue.

Files now go through 4 subdirectories under /tmp/batch before being removed.

jobs: files are created and placed here by mbatch_process
farm: the batcher.tcl server moves the file to be processed here before
      telling the client to use it.  this removes the conflict that
      occurred when a second client would request a file before the
      first client had moved it to work.
work: the client moves the file to work when processing is started
done: the client moves the file to done when processing is completed.

mbatch_process() (using show_tiles.pl) looks in the done directory to
see which files have been completed, updates the map display, and then
removes the files.

To run the batch server:
  ./batcher.tcl server

To run a client (up to 1 per cpu or core available):
  ./batcher.tcl localhost

By running two clients, processing time for batch_process can be cut in half.