Changes made to functions in veg.i:
There was a weird "bug" in the processing of first return and bare earth elevation which occured during during flightline turns and along the edges of the raster scan.  The first surface algorithm assigned a "null" value to those pulses that could not be assigned a valid first return elevation value. This "null" value was < -1000 m -- an obvious outlier -- however, the laser pulses with these outliers remained in the data even after "Test and Clean" was invoked, and often remained in the data even after RCF was invoked.  We never saw the outliers when plotting data because the colorbar range was never big enough to include these outliers.  Worse, the outliers had incorrect easting and northing values as well.

During the December holidays, I worked extensively on the bare earth algorithm, and instead of assigning -1000 as null value, I decided to assign the mirror elevation to a potential outlier.  The bare earth plots showed these outliers in the data when the color bar range extended beyond the mirror elevation values.

The algorithms are now fixed --- the east, north and elevation values are assigned to 0 for any outliers, which are easily removed by the 'Test and Clean' function.

This "bug" should not have affected any prior data processing, except that some outliers may have remained in the data unless removed during manual filtering.