bottom_return.i geo_bath.i surface_topo.i veg.i
When creating edf files, no primitives were explicitly set which led to the
binary files being written using the current platform's word sizes for various
data types. This was causing problems because a long is 4 bytes on 32-bit
machines but is 8 bytes on 64-bit machines.

I've corrected this by explicitly setting the binary files to use i86
primitives. This forces the 64-bit machine to use the same type sizes as the
32-bit machines.