initialize_cir.py
This script initializes a raw CIR dataset. For each mission day that has CIR
imagery and trajectories, it extracts all CIR images and copies the
trajectories. It can use both a local directory or a remote directory (over
SSH) as a source directory.

This basically accomplishes the same thing as makecir.pl + fixup.pl, except it
does it in a single pass for an entire mission and additionally grabs the
trajectories as well.