Removed CVS keywords from all files in lidar-processing/src (but did not check files in subdirectories).
There are two primary reasons for this change:

1. When sourcing files like eaarl.i and l1pro.i, so much $Id$ information was
getting dumped to the screen that it could obscure error messages or other
important information. Meanwhile, the keywords provide no information that
isn't easily available by just asking CVS for the information.

2. Keywords are incompatible (or less supported/meaningful) in other revision
control systems, and would have to be removed during migration if/when we
migrate to another revision control system.