read_yfile.i - read_ascii_xyz
Function read_ascii_xyz has been *completely* changed. You can no longer use it
for the same purposes it was formerly used for.

Formerly, the function read simple numerical files and returned an array of
numbers, similar to read_ascii. This functionality is now handled by read_ascii
from Yutils.

The new read_ascii_xyz now does a reverse of write_ascii_xyz: it reads an ASCII
xyz file and returns an ALPS data structure holding the data. It is optimized
to detect and parse the output of write_ascii_xyz, but is generalizable enough
that it can likely be used in other scenarios as well. The function is well
documented; see its help for more info.