general.i Added func pbd_load. Instead of writing this:     f = openb(filename);     data = get_member(f, f.vname);     close, f; You can now write this:     data = pbd_load(filename);
Beyond just returning the data, the pbd_load also does a wide range of error
checking to avoid having Yorick crash when something doesn't go as expected. If
anything goes wrong, it will simply return []. The error message, if any, will
be received in an output variable. The file's vname will also be returned in an
output variable. For example:
    data = pbd_load(filename, err, vname);
If err is nil, then the data loaded properly and vname is the data's vname.
Otherwise, err contains a string specifying what error was encountered. This
allows a calling function to detect and handle a wide range of errors safely
without having to roll the same code up each time.

See help, pbd_load for more details.