dirload.i This new file freshly implements the ability to load a directory of PBD data. Its functionality is meant to supercede (and consolidate) these functions:   merge_data_pbds (in read_yfile.i)   zoneload_dt_dir (in zone.i)   zoneload_qq_dir (in zone.i)   sel_rgn_from_datatiles (in data_rgn_selector.i)
This is acheived through the function dirload. This function is, in its
simplest form, the same as merge_data_pbds. However, it has the special feature
of allowing the user to pass filters to it that can alter which files get
loaded and which data gets loaded from those files.

These filters are not intended to be created by the user directly, rather,
there are several functions that create them for the user:
  dlfilter_rezone (coerces all data to a specific zone)
  dlfilter_poly (only loads data within a defined polygon)
  dlfilter_bbox (only loads data within a defined bounding box)

GUI for this functionality will be committed soon.