l1pro.ytk Added the following menu entries. * Utilities -> Launch statistics by... -> Flightline * Utilities -> Launch statistics by... -> Flightline and digitizer * Utilities -> Launch statistics by... -> Day * Utilities -> Launch statistics by... -> Day and digitizer
The above menu entries are similar to what can be achieved by using "Launch
segments by..." and then clicking the "Statistics" button. However, these new
menu entries skip the data segments GUI and do not add the variables to the
varlist.

Thus, the variable names you see in the statistics GUI when using this will NOT
exist (unless you created them on your own). They are just names. Conversely, if you later launch the data segments GUI, do NOT count on the variables to match up.