mission_conf.i mission_conf.ytk mosaic_tools.i
Updated the Mission Configuration tool to refer to the INS data as "INS"
instead of "DMARS", both internally and externally. Also updated code that was
impacted by this.

In the Mission Configuration tool, added a menu entry Actions -> Generate KMZ.
This makes it simpler to generate KMZ files for a mission.