Adding README to explain the purpose of the attic. Here's its contents:
This attic is a place where unused code can get shuffled away for safe storage.

Ideally, files that get added or moved here should be prefixed with YYYY-MM- to
indicate when they were moved in. This will help other users get a sense for
which files they might want to review if they're coming back to ALPS after a
while and want to know what's been pushed out of sight.

Ideally, files should also receive some commentary at the top explaining when
they were moved and why. Why are we no longer using this code? Was it replaced
by some other set of code? Did we abandon whatever we were trying to do in it?
Was it just a temporarily thrown-together script that isn't really useful in
the long term?

If for some reason you'd like to use some code that's located in the attic, you
can #include it:
  #include "attic/2010-01-example.i"
Of course, you shouldn't do this in code you're planning to commit. If you're
planning to commit code that refers to something in the attic, that's a sign
that either the code you're writing should be updated (to reflect the changes
that led to the file going in the attic) or else the code in the attic should
be dusted off and put back in active use.