edb_access.i Previously, the location of the TLD files was hard-coded to expect them in the "/eaarl/" subdirectory of the data path. However, we sometimes have missions that end up with eaarl-am, eaarl-pm, etc. that split those files up. Previously, we had to either juggle symlinks or split the data up into seperate missionday directories for it to load.
This has changed: The TLD files are now expected to be in the same directory as
your EDB index file (edb_filename). Thus, if the file is in an eaarl
subdirectory of data_path, it works as it did before. However, it now also
works if the file is in any arbitrary directory as well -- provided the index
file is kept with the TLD files.