Added C implemented functions to C-ALPS to convert between geographic and UTM coordinates.
Also added a function calps_compatibility that should be useful if, in the
future, any changes ever get made within C-ALPS that break backwards
compatibility. This function will give the Yorick code in lidar-processing/src
a means of determining whether the functions in the available C-ALPS is
incompatible with what it expects, should such a situation arise. At the
moment, however, it is not needed.