Major changes to NAVD88 offset code. There are now four different interpolation routines that can be used when calculating the offsets:   n88_interp_spline2d   n88_interp_qfit2d   calps_n88_interp_spline2d   calps_n88_interp_qfit2d The datum conversion routines will use function n88_interp2d by default. This function is a wrapper that automatically uses calps_n88_interp_spline2d if C-ALPS is available, or n88_interp_qfit2d if it is not.
The spline2d algorithms use two-dimensional spline fit interpolation with a 6x6
neighborhood (downgraded to 4x4 or 2x2 if near the edges of the grid). Both
versions use the same internal logic. This algorithm is equivalent to the
spline interpolation algorithm used in the utility programs released with
recent GEOID models.

The qfit2d algorithms use two-dimensional parabolic interpolation with a 3x3
neighborhood. Both versions use the same internal logic. The previous versions
of this file used the parabolic interpolation, which is based on the code for
the utility programs provided with the GEOID models from the 90's.

The versions prefixed by calps_ are implemented in C and run much, much faster
than the other versions, whihc are implemented in Yorick. In an informal test
that interpolated approximately 100 million points, the four functions had
these run times:
  n88_interp_qfit2d - 6 seconds
  n88_interp_spline2d - 1 minute, 10 seconds
  calps_n88_interp_qfit2d - 1 minute, 32 seconds
  calps_n88_interp_spline2d - 22 minutes, 39 seconds

However, this is only a component of the work done during a normal datum
conversion. In an informal test that batch datum converted 130 files (3 GB)
from WGS84 to NAVD88, the four functions resulted in these overall run times:
  n88_interp_qfit2d - 8 minutes, 58 seconds
  n88_interp_spline2d - 10 minutes, 34 seconds
  calps_n88_interp_qfit2d - 10 minutes, 44 seconds
  calps_n88_interp_spline2d - 39 minutes, 21 seconds
When run using the former revision of this file, the run time was 10 minutes, 7
seconds.

Additionally, the results of the above batch datum conversions were compared.
The results observed from all four functions agreed to within +/- 1cm of the
results seen under the previous revision, and with each other.

Also added a new function import_geoid_grid. This imports a geoid model into a
ZGRID structure, allowing it to be visualized more easily.