*** NOTE: This is a significant modification to the bathymetry algorithm.  It is possible that this algorithm will change the results of processing bathymetry waveforms.  If you encounter any serious problems or see very different (incorrect) results, please contact Amar (anayegandhi@usgs.gov) immediately. Rigorous testing of this new approach is currently underway in St. Pete. ***
Modified function ex_bath to select the "last" peak in the bathymetric waveform (after correcting for the exponential decay of the laser pulse through the water column). Before this mod, the bathy algorithm would select the most significant peak (often the 2nd peak after the surface return) as the bottom return. This change to the algorithm now ensures that the "last" peak above a pre-defined threshold is selected.  This will (should) greatly improve the ability to process waveforms that include significant returns from turbidity in the water column.

The remove_noisy_tail() function is applied to the resulting bathy waveform.
The function uses a new "peak detection" method (as defined in extract_peaks_first_deriv() in wf_analysis.i) to find the significant inflections (peaks) in the waveform.